added moving functions to numbers
